Split out optionalField.ts
