+++ experimental/dds/tree2/src/feature-libraries/default-field-kinds/index.ts
@@ -5,7 +5,6 @@
 
 export {
 	FieldKinds,
-	BrandedFieldKind,
 	ValueFieldKind,
 	Optional,
 	Sequence,
