+++ experimental/dds/tree2/src/feature-libraries/sequence-field/fence.json
@@ -1,5 +1,5 @@
 {
 	"tags": ["sequence-field"],
 	"exports": ["index"],
-	"imports": ["codec", "core", "feature-libraries", "util"]
+	"imports": ["codec", "core", "feature-libraries", "util", "modular-schema"]
 }
